Add files via upload
The present code Zodiac is a pseudospectral DNS code, presently configured with two periodic directions. Zodiac is the main code that calls channel.f solver. boundary.f file deals with the estimations of boundary layer fluxes. The user had the choice to ouput both in netcdf as well as vtk format for paraview visualization.